Fixed potential NULL pointer issue
If create_temp_file() returns NULL, this strlen() check would cause
a SEGV.

Acked-by: Gert Doering <gert@greenie.muc.de>

+++ ssl.c
@@ -3194,7 +3194,7 @@ verify_user_pass_script (struct tls_session *session, const struct user_pass *up
     }
 
  done:
-  if (strlen (tmp_file) > 0)
+  if (tmp_file && strlen (tmp_file) > 0)
     delete_file (tmp_file);
 
   argv_reset (&argv);
